Added GameState and small general fixes.
Added a new script called 'GameState.cs'.  This script will be - more or
less the front end of the first scene.  The goal of the script is to:
Allow tutorials to be activated in the start of the game, and give the
player an opportunity to restart the game when they reach a 'game over'
state.
When the correct minion with the 'ID' reaches the exit, the quadratic
equation was never regenerated.
Removed some superfluous code.
Deactivated 'GameController.cs' -- Not enough plans to keep this script
alive yet.
When the score on the HUD is being updated, it will now be done in its
own individual function is 'Score.cs' instead of being shared when the
score is being incremented.  The reason for this change is to allow the
HUD to be updated without the restriction of incrementing only.
Added a new input key 'R' into the Input Manager.  This key is assigned
to restart the game once the 'game over' state has been reached.