Fixed collision confusion
Fixed a bug that would cause the minion actors to miss collision of the
Forward Enabler gameobject.  Originally, when the actor hits the Forward
Enabler, they would re-detect that they are still colliding with the
ladder.  When the minion actors reach this confusion, they would
continue to increment in the y-axis endlessly [demonstration as seen
here: http://youtu.be/4GE7QYc9HfM ].  This fix merely changes the height
of the forward enabler on the Y and Z axis, and enlarging the exit
gameobject for assurance sakes.