Documentation Revision && Music Enabled
In the last revision, I turned off the music and forgot to turn it back
on before I committed the new revision.  I simply just turned it back
on.
Added documentation to the scripts.  This has been done several times in
the past, but all of the changes have been lost -- this was when we used
Microsoft OneDrive, and we couldn't simply synchronize our work
together.