Code Revision: LetterBox
Slightly revised the code structure.
No longer is it required for the variable to cache the integer from the
RNG to be compared within the Switch statement, instead just directly
use the RNG within the Switch's condition.
Generate function is now private, thus outside classes\scripts must use
Access_Generate() to invoke the Generate() function.  This is done to
assure if there is any internal code changes, the only alterations
needed will be within the class itself and not the entire game's
scripts.