Very General Revision and Bug Fixes
Added Win Lost text
When the user lost or won the game, the minions were not cleared from
the scene
Fixed a comparison bug in Driver(); it would not compare the Minion and
the generated equation index.
When the game is restart, the variables 'gameOverFail' and 'gameOverWin'
will be set to false.
When a minion triggers the collider, the collider function will send a
signal to the GameEvent script to activate for comparison of the
minion's ID and the equation index.
Updated the scene