Moved Branch 'MinionMath[...]CodeRevision to MASTER
Despite being out of practice (I gave up on trying to merge one branch
to another and spent +5 hours on reaching problems after problems with
missing exe's), I just copy and pasted the files that was in the Branch
'MinionMathMayhem_CodeRevision' to the MASTER (or Trunk) branch.

To my knowledge, the game itself is still intact and fully functional.