Fixed bug that caused both minions to be selected
When the minions collide with each other, it was possible that - if the
parent IF wasn't satisfied, the else will force the other actor to be
selected even if it was not supposed to be selected.  Thus, the else
must satisfy a condition in order for it to be true.