Made Cross-Values more vigorous
Do not check 'actor1' <==> 'minionCollision_Actor1'; When the second
actor calls the main collision function, it must be 'actor1' <==>
'minionCollision_Actor2' as its value is stored from the previous
transaction.
If the second actor calls the collision function, check ALL values of
the colliding minions from the previous transaction, not just one.