Added moving number
I added the moving number to the project. I am still working on
implementing it, but I am currently stuck on how to make only part of
the game pause. Wanted to upload this before someone else makes any
changes, since I won't be able to work on it anymore for today.