Optimization improvement; less CPU intensive update
Reduced CPU wasted cycles when the event letter is refreshed and forcing
the RAM to be consistently reset to the same value over several
thousands of times before the next - actual needed update from the game,
due to letter change or correct answer.
GameEvent: Partial revision of documentation, for elements not related
to the optimization update.