Another way to prevent a null, force it as '1'.
Besides using a do-while loop to prevent a zero, just return '1'
instead.  I decided that this is the best route to prevent a
hypothetical situation like: Looping 50 times, all return 0 and the
request is to not return a null value.  Having a loop will gradually
degrade performance and play-time if this situation happens.