Added Options/Rounds
I added 2 more rounds to the treadmill game (12 in total) which ask the
user for x when x is already given (act as introduction rounds). I also
added an option in the pause menu to disable obstacles if the user wants
to focus more on the problem itself. Also changed some in-code stuff for
the gameControl script which makes it more adaptable to other changes.