Added script 'Minion_LifeSpan'
This script is designed to manage the actor's life span.  To establish
it's death, another script must call a function within this script for
the value to be initialized.  Then when wanting to fetch it's lifespan,
just call the access function.
It is also possible to determine if the actor is still 'alive'.  But,
this only determines by the timeOfDeath var.  If the var is Null the
actor is alive, but any other value means that the actor is dead.