Added Wave Delay and Fixed Bugs
I fixed some bugs with the spawning of the barrels and also added a
delay to the wave. This delay takes effect if the player collides with
something. Once he collides with an object the spawning of obstacles
stops and waits for 5 seconds until it starts again. This is meant to
give the player a short break.