Expunged a work around for the Daemon Servicer
I added this work around as the Daemon Servicer does not have any
control on the spawner functionality; a new spawner controller will be
made - or just modify the existing and rehash it.
In this revision, if the the Daemon Servicer is activated, the game will
not function properly as the work around has been expunged.