Spawn Code Revised; Good Bye Legacy Extension! :)
Revised the spawn code to work with the daemon service.  The code is
still used from the previous WaveManager(), but revised slightly to be
more condensed.
Expunged Legacy mode!  Now all users must use the new Daemon service,
with no option to opt out.  Though, this daemon functionality is still
under WIP, the benefits should out-weigh the older scheme of ideas back
in the past.