Expunged the SpawnService script.
Despite having already created, it is better to leave it part of the AI
structure and be sure that it will function properly.  It might be
possible to just rename this script and the class, however - I don't
personally want to waste time on making the script work with Unity with
ease -- I remember had issues doing this exact task before.