Bug Fixes and Smart Number Generator
Added the new instructions send by David. Also fixed a bug where the red
danger area would stay after hitting a number at a specific time (it was
pretty hard to get that bug but it happenes). So now the red danger area
gets removed along with other obstacles. Also added a smart number
generator. Now, the numbers on the coins get generated with a min and
max that changes according to what wave the user is on. This way the
range of the number on the coins decreases the longer the user stays on
the same round without getting hit.