GameController now listens AI mastery for tutorial
When the user is not performing well or struggling with the game, the AI
UserMastery will openly broadcast a message to allow the user to view a
demonstration video.  The GameController, which is the spine of the
game, will try to interact with the request and append to a specific
protocol.