[WIP] AI Tutorial sequence
Very minor revisions; renamed function and revised a function header
comment.
When the AI reports that the user needs demonstration help, call the
GameExecute_Tutorial function.  Everything we essentially we need is
already there.  However, this function will need to be restructured.

KNOWN BUGS: When tutorial is playing by request by the AI, the game
(spawners) will not stop.