Revised how the SkipButton works.
To avoid issues with the current code structure, the old ways of how the
SkipButton worked - is now obsolete.   Instead, we have to restructure
the code and the controls via Unity UI Button via Inspector.  Instead of
'OnDisable' which now causes this to be called regardless if the button
was never clicked - to execute it's instruction automatically which
causes issues.  Instead, when clicked, call a specific function and then
it will properly execute when needed and avoid execution when NOT
needed.