Revised when control value is initially set
Changed when the variable 'control' is set to 'True'.  Previously, the
Awake() function will only be called once after the actor has been
activated within the virtual world, but will not be called again.
Instead, as we need this object to be activated again and this variable
must be updated, I moved this statement to the 'OnEnable()' function.