Allow the user to further evaluate the same equation
It is now possible to allow the user to continue evaluating the same
generated equation.  For example, if the user gets a correct answer, the
equation will not be regenerated until the user gets another correct
answer.  This can be changed using a new short variable "Skip Generate
Equation Threshold".  Default value is 1