Fixed GameManager() issues
Finally fixed this issue [
https://github.com/joseffaghihi/GamificationAlgebraConcepts/issues/25 ],
turns out to be a massive over-sight....  This function is no longer
killed, as it does its own waiting, which is also fixed now.
When the Tutorial is running, the GameManager() will now actually wait
for the tutorial to finish, previously - it would call IEnumator and
move on without waiting for the IENumator to close properly.