Overall grading system is now updated properly
The tentative grading system will now properly adjust the overall
grading system as needed.  This sort of dependency is needed as the
tentative score can weed out scores that only reflects on the
interactive tutorials or the like.  The overall grading system should
only capture the scores from the main game.