Resurrected grade checking and game challenging
The user's score is now checked and evaulated.  This change provides
flexibilty for the DEG to activate if the user is performing well.  But,
also provides a tutorial if the user is not doing so well with the
material.

This feature is still not entirely stable; still WIP at the moment.